Move post processor to avoid circular import.
