+++ safe/definitions/__init__.py
@@ -14,8 +14,8 @@ from safe.definitions.constants import *
 from safe.definitions.default_values import *
 from safe.definitions.exposure import *
 from safe.definitions.exposure_classifications import *
-from safe.definitions.fields import *
 from safe.definitions.field_groups import *
+from safe.definitions.fields import *
 from safe.definitions.hazard import *
 from safe.definitions.hazard_category import *
 from safe.definitions.hazard_classifications import *
@@ -25,9 +25,7 @@ from safe.definitions.layer_modes import *
 from safe.definitions.layer_purposes import *
 from safe.definitions.messages import *
 from safe.definitions.minimum_needs import *
-from safe.definitions.post_processors import *
 from safe.definitions.provenance import *
+from safe.definitions.reports import *
 from safe.definitions.units import *
 from safe.definitions.versions import *
-from safe.definitions.reports import *
-from safe.definitions.post_processors import *
