+++ safe/gis/vector/summary_1_aggregate_hazard.py
@@ -7,6 +7,7 @@ import logging
 from PyQt4.QtCore import QPyNullVariant
 from qgis.core import QGis, QgsFeatureRequest
 
+from safe.post_processors import post_processor_affected_function
 from safe.definitions.fields import (
     aggregation_id_field,
     aggregation_name_field,
@@ -22,7 +23,6 @@ from safe.definitions.fields import (
 from safe.definitions.hazard_classifications import not_exposed_class
 from safe.definitions.layer_purposes import (
     layer_purpose_aggregate_hazard_impacted)
-from safe.definitions.post_processors import post_processor_affected_function
 from safe.definitions.utilities import definition
 from safe.gis.sanity_check import check_layer
 from safe.gis.vector.summary_tools import (
