+++ safe/post_processors/__init__.py
@@ -1,13 +1,14 @@
 # coding=utf-8
-from safe.definitions.post_processors.productivity_post_processors import \
-    productivity_post_processors
-from minimum_needs_post_processors import *
-from population_post_processors import *
-from post_processor_functions import *
-from post_processor_inputs import *
-from post_processors import *
-from safe.definitions.post_processors.minimum_needs_post_processors import \
+from safe.post_processors.minimum_needs_post_processors import \
     minimum_needs_post_processors
+
+from minimum_needs_post_processors import *
+from safe.post_processors.population_post_processors import *
+from safe.post_processors.post_processor_functions import *
+from safe.post_processors.post_processor_inputs import *
+from safe.post_processors.post_processors import *
+from safe.post_processors.productivity_post_processors import \
+    productivity_post_processors
 
 # This is the order of execution, so the order is important.
 # For instance, the size post processor must run before size_rate.
