+++ safe/post_processors/minimum_needs_post_processors.py
@@ -2,19 +2,19 @@
 
 """Postprocessors about additional items in minimum needs."""
 
+from safe.post_processors import (
+    constant_input_type,
+    field_input_type,
+    needs_profile_input_type)
+from safe.post_processors import (
+    formula_process,
+    multiply,
+    function_process)
 from safe.definitions import minimum_needs_fields, displaced_field
 from safe.definitions.concepts import concepts
 from safe.definitions.fields import additional_rice_count_field
 from safe.definitions.fields import (
     pregnant_displaced_count_field, lactating_displaced_count_field)
-from safe.definitions.post_processors.post_processor_inputs import (
-    constant_input_type,
-    field_input_type,
-    needs_profile_input_type)
-from safe.definitions.post_processors.post_processors import (
-    formula_process,
-    multiply,
-    function_process)
 from safe.utilities.i18n import tr
 
 __copyright__ = "Copyright 2017, The InaSAFE Project"
